lpc176x: chipid as usb serial number cleanup.
