+++ src/generic/usb_cdc.h
@@ -22,7 +22,7 @@ void usb_stall_ep0(void);
 void usb_set_address(uint_fast8_t addr);
 void usb_set_configure(void);
 void usb_request_bootloader(void);
-void usb_set_serial(uint8_t *);
+struct usb_string_descriptor *usbserial_get_serialid(void);
 
 // usb_cdc.c
 void usb_notify_bulk_in(void);
