--- src/lpc176x/uid.h
@@ -1,8 +1,0 @@
-#ifndef __LPC176X_UID_H
-#define __LPC176X_UID_H
-
-#include "stdint.h"
-
-void platform_get_uid(uint32_t *);
-
-#endif // uid.h
